requirements.txt
flask
pymongo
paho-mqtt
gunicorn
